// apps/simulator-web/src/types/scenario.ts
export type ScenarioId =
  | 'NORMAL_COLD_START'
  | 'NORMAL_HIGHWAY_CRUISE'
  | 'TRAFFIC_IDLE'
  | 'HARD_ACCELERATION'
  | 'LOW_BATTERY'
  | 'ALTERNATOR_FAILURE'
  | 'ENGINE_OVERHEATING'
  | 'LEAN_MIXTURE'
  | 'VACUUM_LEAK_PATTERN'
  | 'SENSOR_FAILURE'
  | 'OBD_DISCONNECT'
  | 'GPS_LOSS'
  | 'STORED_DTC'
  | 'CRITICAL_ENGINE_WARNING'
  | 'TURBO_CRUISE'
  | 'TURBO_PULL'
  | 'OVERBOOST'
  | 'GPS_DRIVE'
  | 'GPS_SIGNAL_LOSS'
  | 'NAVIGATION_TURN'
  | 'MUSIC_PLAYBACK'
  | 'MUSIC_WARNING'
  | 'NIGHT_DRIVE'
  | 'PARK_VEHICLE'
  | 'CLASSIC_CRUISE'
  | 'CITY_NAV'
  | 'HIGHWAY_NAV'
  | 'PHONE_NAV_LOSS'
  | 'EXPEDITION_START'
  | 'TRAIL_DRIVE'
  | 'STEEP_CLIMB'
  | 'SIDE_SLOPE'
  | 'WAYPOINT_APPROACH'
  | 'RETURN_TO_START'
  | 'LOW_FUEL_OFFROAD'
  | 'WINCH_OPERATION'
  | 'RALLY_RAID'